backend/requirements.txt
# ============================================================
# OnlyStudents Backend - Python Requirements
# ============================================================

# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database ORM & driver
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Pydantic (data validation)
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Authentication & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Redis (caching / sessions)
redis==5.0.4
aioredis==2.0.1

# AI / Embeddings
sentence-transformers==2.7.0
numpy==1.26.4
pgvector==0.2.5

# HTTP client
httpx==0.27.0

# Environment config
python-dotenv==1.0.1

# File uploads / AWS S3
boto3==1.34.95
Pillow==10.3.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1
